move toIntentData function
